refactor(exporter): restructure TXT and EPUB exporter, integrate TextCleaner
