+++ README.md
@@ -12,7 +12,7 @@
   - EPUB (可选包含章节插图)
 - 支持活动广告过滤:
   - [x] 章节标题
-  - [ ] 章节正文
+  - [x] 章节正文
 
 ---
