+++ docs/todo.md
@@ -15,4 +15,3 @@
   - 提供图片压缩选项 (例如 `caesium`)
   - 提供简单 Web/桌面 GUI 前端
   - 提供可定制的导出模板 (例如 `Jinja2`)
-  - 研究并使用起点移动端 API
