+++ src/novel_downloader/cli/__init__.py
@@ -6,8 +6,8 @@ novel_downloader.cli
 This module exposes the CLI entry point.
 """
 
-from .main import cli_main
-
 __all__ = [
     "cli_main",
 ]
+
+from .main import cli_main
