+++ src/novel_downloader/core/__init__.py
@@ -14,17 +14,6 @@ downloading and processing online novel content, including:
 - Exporter: Responsible for exporting downloaded data into various output formats.
 """
 
-from .downloaders import get_downloader
-from .exporters import get_exporter
-from .fetchers import get_fetcher
-from .interfaces import (
-    DownloaderProtocol,
-    ExporterProtocol,
-    FetcherProtocol,
-    ParserProtocol,
-)
-from .parsers import get_parser
-
 __all__ = [
     "get_downloader",
     "get_exporter",
@@ -35,3 +24,14 @@ __all__ = [
     "FetcherProtocol",
     "ParserProtocol",
 ]
+
+from .downloaders import get_downloader
+from .exporters import get_exporter
+from .fetchers import get_fetcher
+from .interfaces import (
+    DownloaderProtocol,
+    ExporterProtocol,
+    FetcherProtocol,
+    ParserProtocol,
+)
+from .parsers import get_parser
