+++ src/novel_downloader/core/exporters/biquge.py
@@ -5,6 +5,8 @@ novel_downloader.core.exporters.biquge
 
 """
 
+__all__ = ["BiqugeExporter"]
+
 from novel_downloader.core.exporters.registry import register_exporter
 from novel_downloader.models import ExporterConfig
 
@@ -22,6 +24,3 @@ class BiqugeExporter(CommonExporter):
             site="biquge",
             chap_folders=["chapters"],
         )
-
-
-__all__ = ["BiqugeExporter"]
