+++ src/novel_downloader/core/exporters/linovelib/txt.py
@@ -3,11 +3,8 @@
 novel_downloader.core.exporters.linovelib.txt
 ---------------------------------------------
 
-Contains the logic for exporting novel content as a single `.txt` file.
-
-This module defines `linovelib_export_as_txt` function, which assembles and formats
-a novel based on metadata and chapter files found in the raw data directory.
-It is intended to be used by `LinovelibExporter` as part of the save/export process.
+Defines `linovelib_export_as_txt` to assemble and export a Linovelib novel
+into a single `.txt` file. Intended for use by `LinovelibExporter`.
 """
 
 from __future__ import annotations
@@ -15,8 +12,11 @@ from __future__ import annotations
 import json
 from typing import TYPE_CHECKING
 
-from novel_downloader.core.exporters.txt_util import format_chapter
-from novel_downloader.utils import save_as_txt
+from novel_downloader.core.exporters.txt_util import (
+    build_txt_chapter,
+    build_txt_header,
+)
+from novel_downloader.utils import get_cleaner, save_as_txt
 
 if TYPE_CHECKING:
     from .main_exporter import LinovelibExporter
@@ -27,22 +27,30 @@ def linovelib_export_as_txt(
     book_id: str,
 ) -> None:
     """
-    将 save_path 文件夹中该小说的所有章节 json 文件合并保存为一个完整的 txt 文件,
-    并保存到 out_path 下
+    Merge all chapter JSON files under `raw_data/<book_id>` into one TXT,
+    then write it to the exporter's output directory.
 
-    处理流程：
-      1. 从 book_info.json 中加载书籍信息 (包含书名、作者、简介及卷章节列表)
-      2. 遍历各卷, 每个卷先追加卷标题, 然后依次追加该卷下各章节的标题和内容
-      3. 将书籍元信息 (书名、作者、原文截至、内容简介) 与所有章节内容拼接
-      4. 将最终结果保存到 out_path 下 (例如：`{book_name}.txt`)
+    Steps:
+      1. Read metadata from `book_info.json`.
+      2. For each volume:
+         - Clean & append the volume title.
+         - Clean & append optional volume intro.
+         - For each chapter: load JSON, clean title & content, then append.
+      3. Build a header block with metadata.
+      4. Concatenate header + all chapter blocks, then save as `{book_name}.txt`.
 
-    :param book_id: Identifier of the novel (used as subdirectory name).
+    :param exporter: The LinovelibExporter instance.
+    :param book_id: Identifier of the novel (subdirectory under raw data).
     """
     TAG = "[exporter]"
     # --- Paths & options ---
     raw_base = exporter._raw_data_dir / book_id
     out_dir = exporter.output_dir
     out_dir.mkdir(parents=True, exist_ok=True)
+    cleaner = get_cleaner(
+        enabled=exporter._config.clean_text,
+        config=exporter._config.cleaner_cfg,
+    )
 
     # --- Load book_info.json ---
     info_path = raw_base / "book_info.json"
@@ -55,26 +63,27 @@ def linovelib_export_as_txt(
 
     # --- Compile chapters ---
     parts: list[str] = []
-    volumes = book_info.get("volumes", [])
 
-    for vol in volumes:
-        vol_name = vol.get("volume_name", "").strip()
-        vol_intro = vol.get("volume_intro", "").strip()
-        if vol_name:
-            volume_header = f"\n\n{'=' * 6} {vol_name} {'=' * 6}\n\n"
-            parts.append(volume_header)
-            exporter.logger.info("%s Processing volume: %s", TAG, vol_name)
+    for vol in book_info.get("volumes", []):
+        vol_title = cleaner.clean_title(vol.get("volume_name", ""))
+        if vol_title:
+            parts.append(f"\n\n{'=' * 6} {vol_title} {'=' * 6}\n\n")
+            exporter.logger.info("%s Processing volume: %s", TAG, vol_title)
+
+        vol_intro = cleaner.clean_content(vol.get("volume_intro", ""))
         if vol_intro:
             parts.append(f"{vol_intro}\n\n")
+
         for chap in vol.get("chapters", []):
             chap_id = chap.get("chapterId")
-            chap_title = chap.get("title", "")
             if not chap_id:
                 exporter.logger.warning("%s Missing chapterId, skipping: %s", TAG, chap)
                 continue
 
-            chapter_data = exporter._get_chapter(book_id, chap_id)
-            if not chapter_data:
+            chap_title = cleaner.clean_title(chap.get("title", ""))
+
+            data = exporter._get_chapter(book_id, chap_id)
+            if not data:
                 exporter.logger.info(
                     "%s Missing chapter file in: %s (%s), skipping.",
                     TAG,
@@ -84,10 +93,10 @@ def linovelib_export_as_txt(
                 continue
 
             # Extract structured fields
-            title = chapter_data.get("title", chap_title).strip()
-            content = chapter_data.get("content", "").strip()
+            title = cleaner.clean_title(data.get("title", chap_title))
+            content = cleaner.clean_content(data.get("content", ""))
 
-            parts.append(format_chapter(title, content, ""))
+            parts.append(build_txt_chapter(title=title, paragraphs=content, extras={}))
 
     # --- Build header ---
     name = book_info.get("book_name")
@@ -96,21 +105,15 @@ def linovelib_export_as_txt(
     updated = book_info.get("update_time")
     summary = book_info.get("summary")
 
-    fields = [
+    header_fields = [
         ("书名", name),
         ("作者", author),
         ("总字数", words),
         ("更新日期", updated),
+        ("内容简介", summary),
     ]
-    header_lines = [f"{label}: {value}" for label, value in fields if value]
 
-    if summary:
-        header_lines.append("内容简介:")
-        header_lines.append(summary)
-
-    header_lines += ["", "-" * 10, ""]
-
-    header = "\n".join(header_lines)
+    header = build_txt_header(header_fields)
 
     final_text = header + "\n\n" + "\n\n".join(parts).strip()
 
@@ -119,9 +122,9 @@ def linovelib_export_as_txt(
     out_path = out_dir / out_name
 
     # --- Save final text ---
-    try:
-        save_as_txt(content=final_text, filepath=out_path)
+    result = save_as_txt(content=final_text, filepath=out_path)
+    if result:
         exporter.logger.info("%s Novel saved to: %s", TAG, out_path)
-    except Exception as e:
-        exporter.logger.error("%s Failed to save file: %s", TAG, e)
+    else:
+        exporter.logger.error("%s Failed to write novel to %s", TAG, out_path)
     return
