+++ src/novel_downloader/core/exporters/qianbi.py
@@ -5,6 +5,8 @@ novel_downloader.core.exporters.qianbi
 
 """
 
+__all__ = ["QianbiExporter"]
+
 from novel_downloader.core.exporters.registry import register_exporter
 from novel_downloader.models import ExporterConfig
 
@@ -22,6 +24,3 @@ class QianbiExporter(CommonExporter):
             site="qianbi",
             chap_folders=["chapters"],
         )
-
-
-__all__ = ["QianbiExporter"]
