+++ src/novel_downloader/core/fetchers/base/__init__.py
@@ -5,10 +5,10 @@ novel_downloader.core.fetchers.base
 
 """
 
-from .browser import BaseBrowser
-from .session import BaseSession
-
 __all__ = [
     "BaseBrowser",
     "BaseSession",
 ]
+
+from .browser import BaseBrowser
+from .session import BaseSession
