+++ src/novel_downloader/core/fetchers/biquge/__init__.py
@@ -5,10 +5,10 @@ novel_downloader.core.fetchers.biquge
 
 """
 
-from .browser import BiqugeBrowser
-from .session import BiqugeSession
-
 __all__ = [
     "BiqugeBrowser",
     "BiqugeSession",
 ]
+
+from .browser import BiqugeBrowser
+from .session import BiqugeSession
