+++ src/novel_downloader/core/fetchers/esjzone/__init__.py
@@ -5,10 +5,10 @@ novel_downloader.core.fetchers.esjzone
 
 """
 
-from .browser import EsjzoneBrowser
-from .session import EsjzoneSession
-
 __all__ = [
     "EsjzoneBrowser",
     "EsjzoneSession",
 ]
+
+from .browser import EsjzoneBrowser
+from .session import EsjzoneSession
