+++ src/novel_downloader/core/fetchers/linovelib/__init__.py
@@ -5,10 +5,10 @@ novel_downloader.core.fetchers.linovelib
 
 """
 
-from .browser import LinovelibBrowser
-from .session import LinovelibSession
-
 __all__ = [
     "LinovelibBrowser",
     "LinovelibSession",
 ]
+
+from .browser import LinovelibBrowser
+from .session import LinovelibSession
