+++ src/novel_downloader/core/fetchers/qianbi/__init__.py
@@ -5,10 +5,10 @@ novel_downloader.core.fetchers.qianbi
 
 """
 
-from .browser import QianbiBrowser
-from .session import QianbiSession
-
 __all__ = [
     "QianbiBrowser",
     "QianbiSession",
 ]
+
+from .browser import QianbiBrowser
+from .session import QianbiSession
