+++ src/novel_downloader/core/fetchers/sfacg/__init__.py
@@ -5,10 +5,10 @@ novel_downloader.core.fetchers.sfacg
 
 """
 
-from .browser import SfacgBrowser
-from .session import SfacgSession
-
 __all__ = [
     "SfacgBrowser",
     "SfacgSession",
 ]
+
+from .browser import SfacgBrowser
+from .session import SfacgSession
