+++ src/novel_downloader/core/fetchers/yamibo/__init__.py
@@ -5,10 +5,10 @@ novel_downloader.core.fetchers.yamibo
 
 """
 
-from .browser import YamiboBrowser
-from .session import YamiboSession
-
 __all__ = [
     "YamiboBrowser",
     "YamiboSession",
 ]
+
+from .browser import YamiboBrowser
+from .session import YamiboSession
