+++ src/novel_downloader/core/interfaces/__init__.py
@@ -14,14 +14,14 @@ Included protocols:
 - ExporterProtocol
 """
 
-from .downloader import DownloaderProtocol
-from .exporter import ExporterProtocol
-from .fetcher import FetcherProtocol
-from .parser import ParserProtocol
-
 __all__ = [
     "DownloaderProtocol",
     "ExporterProtocol",
     "FetcherProtocol",
     "ParserProtocol",
 ]
+
+from .downloader import DownloaderProtocol
+from .exporter import ExporterProtocol
+from .fetcher import FetcherProtocol
+from .parser import ParserProtocol
