+++ src/novel_downloader/core/parsers/esjzone/__init__.py
@@ -5,6 +5,6 @@ novel_downloader.core.parsers.esjzone
 
 """
 
-from .main_parser import EsjzoneParser
-
 __all__ = ["EsjzoneParser"]
+
+from .main_parser import EsjzoneParser
