+++ src/novel_downloader/core/parsers/linovelib/__init__.py
@@ -5,6 +5,6 @@ novel_downloader.core.parsers.linovelib
 
 """
 
-from .main_parser import LinovelibParser
-
 __all__ = ["LinovelibParser"]
+
+from .main_parser import LinovelibParser
