+++ src/novel_downloader/core/parsers/qianbi/__init__.py
@@ -5,6 +5,6 @@ novel_downloader.core.parsers.qianbi
 
 """
 
-from .main_parser import QianbiParser
-
 __all__ = ["QianbiParser"]
+
+from .main_parser import QianbiParser
