+++ src/novel_downloader/core/parsers/yamibo/__init__.py
@@ -5,6 +5,6 @@ novel_downloader.core.parsers.yamibo
 
 """
 
-from .main_parser import YamiboParser
-
 __all__ = ["YamiboParser"]
+
+from .main_parser import YamiboParser
