+++ src/novel_downloader/utils/text_utils/__init__.py
@@ -6,13 +6,15 @@ novel_downloader.utils.text_utils
 Utility modules for text formatting, font mapping, cleaning, and diff display.
 
 Submodules:
-- font_mapping: Replace obfuscated characters using font maps
-- chapter_formatting: Build structured chapter strings from raw content
-- text_cleaning: Remove promo text and check for spam lines
 - diff_display: Generate inline diffs with aligned character markers
+- numeric_conversion: Convert between Chinese and Arabic numerals
+- text_cleaner: Text cleaning and normalization utilities
+- truncate_utils: Text truncation and content prefix generation
 """
 
 __all__ = [
+    "TextCleaner",
+    "get_cleaner",
     "content_prefix",
     "truncate_half_lines",
     "chinese_to_arabic",
@@ -25,6 +27,7 @@ from .numeric_conversion import (
     arabic_to_chinese,
     chinese_to_arabic,
 )
+from .text_cleaner import TextCleaner, get_cleaner
 from .truncate_utils import (
     content_prefix,
     truncate_half_lines,
