+++ src/novel_downloader/utils/time_utils/__init__.py
@@ -12,11 +12,11 @@ Includes:
     Sleeps for a random duration, useful for human-like delays or rate limiting.
 """
 
-from .datetime_utils import calculate_time_difference
-from .sleep_utils import async_sleep_with_random_delay, sleep_with_random_delay
-
 __all__ = [
     "calculate_time_difference",
     "async_sleep_with_random_delay",
     "sleep_with_random_delay",
 ]
+
+from .datetime_utils import calculate_time_difference
+from .sleep_utils import async_sleep_with_random_delay, sleep_with_random_delay
